FREESCAPE: Add explicit float cast to make MSVC happy

+++ engines/freescape/gfx_opengl.cpp
@@ -299,7 +299,7 @@ void OpenGLRenderer::renderPlayerShootBall(byte color, const Common::Point posit
 
 	glColor3ub(r, g, b);
 	int triangleAmount = 20;
-	float twicePi = 2.f * M_PI;
+	float twicePi = (float)(2.0 * M_PI);
 	float coef = (9 - frame) / 9.0;
 	float radius = (1 - coef) * 4.0;
